tests: move e2e test_blocks
